+++ lib/src/in_app_webview/in_app_webview.dart
@@ -76,6 +76,7 @@ class InAppWebView extends StatefulWidget implements WebView {
     this.onTitleChanged,
     this.onWindowFocus,
     this.onWindowBlur,
+    this.onOverScrolled,
     this.androidOnSafeBrowsingHit,
     this.androidOnPermissionRequest,
     this.androidOnGeolocationPermissionsShowPrompt,
@@ -307,6 +308,9 @@ class InAppWebView extends StatefulWidget implements WebView {
 
   @override
   final void Function(InAppWebViewController controller)? onExitFullscreen;
+
+  @override
+  final void Function(InAppWebViewController controller, int x, int y, bool clampedX, bool clampedY)? onOverScrolled;
 
   @override
   final Future<WebResourceResponse?> Function(